requirements.txt
solana-py==0.30.0
